Extract levenshtein submodule, move proptests to integration tests
Split the Wagner-Fischer distance fn into src/levenshtein.rs alongside its
known-answer test. Promote the contains/fuzzy proptests to a single-binary
tests/integration/ harness (mirroring term_suffix_index), giving the fuzzy
oracle an independent naive-recursive distance so it no longer compares the
crate's levenshtein against itself.
